refactor: naming change(pin -> score)
