+++ src/test/java/bowling/domain/score/PinTest.java
@@ -1,4 +1,4 @@
-package bowling.domain.pin;
+package bowling.domain.score;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
